+++ src/main/java/br/com/senac/ado/zoologico/service/HabitatService.java
@@ -1,6 +1,8 @@
 package br.com.senac.ado.zoologico.service;
 
+import br.com.senac.ado.zoologico.dto.HabitatDTO;
 import br.com.senac.ado.zoologico.entity.Habitat;
+import br.com.senac.ado.zoologico.exception.ResourceNotFoundException;
 import br.com.senac.ado.zoologico.repository.HabitatRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,28 +16,38 @@ public class HabitatService {
 
     private final HabitatRepository repository;
 
-    public List<Habitat> listarTodos() {
+    public List<Habitat> getAll() {
         return repository.findAll();
     }
 
-    public Habitat buscar(UUID id) {
-        return repository.findById(id).orElse(null);
+    public Habitat findById(UUID id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não Encontrado com o ID informado"));
     }
 
-    public Habitat salvar(Habitat habitat) {
-        return repository.save(habitat);
+    public UUID salvar(HabitatDTO dto) {
+
+        Habitat habitat = new Habitat();
+        habitat.setNome(dto.getNome());
+        habitat.setTipo(dto.getTipo());
+        habitat.setAreaM2(Double.valueOf(dto.getAreaM2()));
+
+        return repository.save(habitat).getId();
     }
 
-    public Habitat atualizar(UUID id, Habitat habitat) {
+    public Habitat update(UUID id, HabitatDTO dto) {
         Habitat existente = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Habitat não encontrado"));
-        existente.setNome(habitat.getNome());
-        existente.setTipo(habitat.getTipo());
-        existente.setAreaM2(habitat.getAreaM2());
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não Encontrado com o ID informado"));
+        existente.setNome(dto.getNome());
+        existente.setTipo(dto.getTipo());
+        existente.setAreaM2(Double.valueOf(dto.getAreaM2()));
         return repository.save(existente);
     }
 
-    public void excluir(UUID id) {
+    public void delete(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não Encontrado com o ID informado para exclusão");
+        }
         repository.deleteById(id);
     }
 }
